feat(DTFS2-6182): update tests
