+++ dtfs-central-api/api-tests/api.js
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
 const request = require('supertest');
 const dotenv = require('dotenv');
