+++ trade-finance-manager-api/.eslintrc.js
@@ -20,7 +20,7 @@ module.exports = {
     ],
     'no-console': ['error', { allow: ['info', 'error'] }],
     'no-underscore-dangle': ['error', { allow: ['_id', '_csrf'] }],
-    'import/no-extraneous-dependencies': ['error', { devDependencies: ['**/*.test.js', '**/*.api-test.js', '**/api-tests/**'] }],
+    'import/no-extraneous-dependencies': ['error', { devDependencies: ['**/*.test.js', '**/*.api-test.js', '**/api-tests/**', '**/__mocks__/**'] }],
     'import/no-named-as-default': 'off',
     'implicit-arrow-linebreak': 'off',
     'object-curly-newline': [
