+++ trade-finance-manager-api/api-tests/v1/deals/deals-party-db.api-test.js
@@ -1,3 +1,4 @@
+const { mockUpdateDeal } = require('../../../src/v1/__mocks__/common-api-mocks');
 const { addPartyUrns } = require('../../../src/v1/controllers/deal.party-db');
 
 describe('add partyUrn to deal', () => {
@@ -7,45 +8,66 @@ describe('add partyUrn to deal', () => {
     eligibility: {},
   };
 
+  beforeAll(() => {
+    jest.resetAllMocks();
+  });
+
+  afterEach(() => {
+    jest.resetAllMocks();
+  });
+
   it('should return false when no deal passed as parameter', async () => {
+    mockUpdateDeal();
+
     const noDeal = await addPartyUrns();
     expect(noDeal).toEqual(false);
   });
 
   it('should return false when companies house no is not given', async () => {
+    mockUpdateDeal();
+
     const deal = {
       ...mockDeal,
       exporter: {
         companiesHouseRegistrationNumber: '',
       },
     };
+
     const noCompaniesHouse = await addPartyUrns(deal);
     expect(noCompaniesHouse.tfm.parties.exporter.partyUrn).toEqual('');
   });
 
   it('should return false when companies house no is not matched', async () => {
+    mockUpdateDeal();
+
     const deal = {
       ...mockDeal,
       exporter: {
         companiesHouseRegistrationNumber: 'NO_MATCH',
       },
     };
+
     const noMatch = await addPartyUrns(deal);
     expect(noMatch.tfm.parties.exporter.partyUrn).toEqual('');
   });
 
   it('should return the deal with partyUrn is successfully matched', async () => {
+    mockUpdateDeal();
+
     const deal = {
       ...mockDeal,
       exporter: {
         companiesHouseRegistrationNumber: 'MATCH',
       },
     };
+
     const match = await addPartyUrns(deal);
     expect(match.tfm.parties.exporter.partyUrn).toEqual('testPartyUrn');
   });
 
   it('should retain existing tfm data', async () => {
+    mockUpdateDeal();
+
     const deal = {
       ...mockDeal,
       exporter: {
@@ -55,6 +77,7 @@ describe('add partyUrn to deal', () => {
         mockField: 'mock data',
       },
     };
+
     const match = await addPartyUrns(deal);
     expect(match.tfm).toMatchObject(deal.tfm);
   });
