+++ trade-finance-manager-api/graphql-query-tests/mutation-update-parties.api-test.js
@@ -2,11 +2,11 @@ const { ApolloServer } = require('apollo-server-express');
 const { applyMiddleware } = require('graphql-middleware');
 const { makeExecutableSchema } = require('@graphql-tools/schema');
 const gql = require('graphql-tag');
-
-jest.mock('../src/v1/api');
+const api = require('../src/v1/api');
 
 const typeDefs = require('../src/graphql/schemas');
 const resolvers = require('../src/graphql/resolvers');
+const { mockUpdateDeal } = require('../src/v1/__mocks__/common-api-mocks');
 
 const UPDATE_PARTIES = gql`
   mutation UpdateParties($id: ID!, $partyUpdate: TFMPartiesInput) {
@@ -37,10 +37,18 @@ describe('graphql mutation - update party', () => {
       resolvers,
       schema: schemaWithMiddleware,
     });
+
+    api.updateDeal.mockReset();
+  });
+
+  afterEach(() => {
+    api.updateDeal.mockReset();
   });
 
   describe('before all party URNs complete', () => {
     it('should return updated party details', async () => {
+      mockUpdateDeal();
+
       const partyUpdate = {
         exporter: {
           partyUrn: '111',
