+++ trade-finance-manager-api/graphql-query-tests/mutation-update-probability-of-default.api-test.js
@@ -3,12 +3,13 @@ const { applyMiddleware } = require('graphql-middleware');
 const { makeExecutableSchema } = require('@graphql-tools/schema');
 const gql = require('graphql-tag');
 
-jest.mock('../src/v1/api');
+const api = require('../src/v1/api');
 
 const typeDefs = require('../src/graphql/schemas');
 const resolvers = require('../src/graphql/resolvers');
 
 const MOCK_DEAL = require('../src/v1/__mocks__/mock-deal');
+const { mockUpdateDeal } = require('../src/v1/__mocks__/common-api-mocks');
 
 const UPDATE_PROBABILITY_OF_DEFAULT = gql`
   mutation UpdateProbabilityOfDefault($dealId: ID!, $probabilityOfDefaultUpdate: TFMProbabilityOfDefaultInput) {
@@ -30,9 +31,17 @@ describe('graphql mutation - update probability of default', () => {
       resolvers,
       schema: schemaWithMiddleware,
     });
+
+    api.updateDeal.mockReset();
+  });
+
+  afterEach(() => {
+    api.updateDeal.mockReset();
   });
 
   it('should return updated probability of default', async () => {
+    mockUpdateDeal();
+    
     const mutationVars = {
       dealId: MOCK_DEAL._id,
       probabilityOfDefaultUpdate: {
