+++ trade-finance-manager-api/src/v1/controllers/deal-add-tfm-data/index.api-test.js
@@ -5,13 +5,20 @@ const addDealProduct = require('./dealProduct');
 const addDealPricingAndRisk = require('./dealPricingAndRisk');
 const addDealStage = require('./dealStage');
 const MOCK_DEAL_AIN = require('../../__mocks__/mock-deal');
+const { mockUpdateDeal } = require('../../__mocks__/common-api-mocks');
+const api = require('../../api');
 
 describe('deal submit - add TFM data', () => {
+  afterEach(() => {
+    api.updateDeal.mockReset();
+  });
+
   it('returns an object with results from multiple functions', async () => {
     const mockDeal = mapSubmittedDeal({
       dealSnapshot: MOCK_DEAL_AIN,
       tfm: {},
     });
+    mockUpdateDeal();
 
     const result = await addTfmDealData(mockDeal);
