+++ trade-finance-manager-ui/server/controllers/case/underwriting/lead-underwriter/index.test.js
@@ -27,9 +27,11 @@ const MOCK_USER_TEAM_USER = {
 
 const MOCK_TEAM_UNDERWRITER_MANAGERS = [MOCK_USER_UNDERWRITER_MANAGER];
 
-const session = {
+const TEST_USER_TOKEN = 'test-token';
+
+const SESSION = {
   user: MOCK_USER_UNDERWRITER_MANAGER,
-  userToken: 'test-token',
+  userToken: TEST_USER_TOKEN,
 };
 
 const MOCK_DEAL = {
@@ -46,7 +48,7 @@ const MOCK_DEAL = {
 const dealId = MOCK_DEAL._id;
 
 describe('GET underwriting - lead underwriter', () => {
-  const userCanEdit = userIsInTeam(session.user, [CONSTANTS.TEAMS.UNDERWRITER_MANAGERS, CONSTANTS.TEAMS.UNDERWRITERS]);
+  const userCanEdit = userIsInTeam(SESSION.user, [CONSTANTS.TEAMS.UNDERWRITER_MANAGERS, CONSTANTS.TEAMS.UNDERWRITERS]);
 
   describe('when deal exists', () => {
     const apiGetUserSpy = jest.fn(() => Promise.resolve(MOCK_USER_UNDERWRITER_MANAGER));
@@ -65,7 +67,7 @@ describe('GET underwriting - lead underwriter', () => {
         deal: MOCK_DEAL.dealSnapshot,
         tfm: MOCK_DEAL.tfm,
         dealId: MOCK_DEAL.dealSnapshot._id,
-        user: session.user,
+        user: SESSION.user,
         currentLeadUnderWriter: MOCK_USER_UNDERWRITER_MANAGER,
       });
     });
@@ -105,7 +107,7 @@ describe('GET underwriting - lead underwriter', () => {
         deal: MOCK_DEAL_UNASSIGNED_LEAD_UNDERWRITER.dealSnapshot,
         tfm: MOCK_DEAL_UNASSIGNED_LEAD_UNDERWRITER.tfm,
         dealId: MOCK_DEAL_UNASSIGNED_LEAD_UNDERWRITER.dealSnapshot._id,
-        user: session.user,
+        user: SESSION.user,
         currentLeadUnderWriter: undefined,
       });
     });
@@ -118,7 +120,7 @@ describe('GET underwriting - assign lead underwriter', () => {
       params: {
         _id: dealId,
       },
-      session,
+      session: SESSION,
     };
 
     const getTeamMembersSpy = jest.fn(() => Promise.resolve(MOCK_TEAM_UNDERWRITER_MANAGERS));
@@ -135,8 +137,8 @@ describe('GET underwriting - assign lead underwriter', () => {
       expect(getTeamMembersSpy).toBeCalledTimes(2);
 
       expect(getTeamMembersSpy.mock.calls).toEqual([
-        [CONSTANTS.TEAMS.UNDERWRITER_MANAGERS, session.userToken],
-        [CONSTANTS.TEAMS.UNDERWRITERS, session.userToken],
+        [CONSTANTS.TEAMS.UNDERWRITER_MANAGERS, SESSION.userToken],
+        [CONSTANTS.TEAMS.UNDERWRITERS, SESSION.userToken],
       ]);
     });
 
@@ -146,14 +148,14 @@ describe('GET underwriting - assign lead underwriter', () => {
       // NOTE: api.getTeamMembers stub only returns one team.
       const alphabeticalTeamMembers = sortArrayOfObjectsAlphabetically(MOCK_TEAM_UNDERWRITER_MANAGERS, 'firstName');
 
-      const expectedAssignToSelectOptions = mapAssignToSelectOptions(MOCK_DEAL.tfm.leadUnderwriter, session.user, alphabeticalTeamMembers);
+      const expectedAssignToSelectOptions = mapAssignToSelectOptions(MOCK_DEAL.tfm.leadUnderwriter, SESSION.user, alphabeticalTeamMembers);
 
       expect(res.render).toHaveBeenCalledWith('case/underwriting/lead-underwriter/assign-lead-underwriter.njk', {
         activeSubNavigation: 'underwriting',
         deal: MOCK_DEAL.dealSnapshot,
         tfm: MOCK_DEAL.tfm,
         dealId: MOCK_DEAL.dealSnapshot._id,
-        user: session.user,
+        user: SESSION.user,
         assignToSelectOptions: expectedAssignToSelectOptions,
       });
     });
@@ -186,7 +188,7 @@ describe('GET underwriting - assign lead underwriter', () => {
         params: {
           _id: '1',
         },
-        session,
+        session: SESSION,
       };
 
       await underwriterLeadUnderwriterController.getAssignLeadUnderwriter(req, res);
@@ -211,7 +213,7 @@ describe('POST underwriting - assign lead underwriter', () => {
       params: {
         _id: dealId,
       },
-      session,
+      session: SESSION,
       body: {
         assignedTo: '123-test',
       },
@@ -223,7 +225,7 @@ describe('POST underwriting - assign lead underwriter', () => {
       userId: req.body.assignedTo,
     };
 
-    expect(apiUpdateSpy).toHaveBeenCalledWith({ dealId: MOCK_DEAL._id, leadUnderwriterUpdate: expectedUpdateObj, token: undefined });
+    expect(apiUpdateSpy).toHaveBeenCalledWith({ dealId: MOCK_DEAL._id, leadUnderwriterUpdate: expectedUpdateObj, token: TEST_USER_TOKEN });
   });
 
   describe('when user cannot edit (i.e, NOT in UNDERWRITER_MANAGERS team)', () => {
@@ -253,7 +255,7 @@ describe('POST underwriting - assign lead underwriter', () => {
         params: {
           _id: '1',
         },
-        session,
+        session: SESSION,
       };
 
       await underwriterLeadUnderwriterController.postAssignLeadUnderwriter(req, res);
